Initial support for Swagger 2.0 WIP
This provides an adapter for parsing Swagger 2.0 documents. It is currently
incomplete as it does not provide full support for paramters. It also needs
more testing to ensure it correctly defines data, specifically schemas.

Additionally, this is relying on the current API Description and Resource
Description namespaces of Refract, which may be changing soon. I have left the
code fairly verbose and non-modular at this point so I can easily change it if
we make changes.