Prevent bug overwriting resource href template.
This change prevents an issue where multiple transitions on a resource could
lead to the resource `href` being overwritten as each transition is processed.
The error occurs when transitions define href variables (params in Swagger).

I believe the correct behavior is to set an `href` on the resource based on
the resource href variables, then for each transition if it has a different
`href` it will get set on the transition.