fix(contributing): Never create venv inside docs
If the venv is created inside docs/, then all the dependencies will be
installed on docs directory including Sphinx itself. When we try to
build docs when there is Sphinx source inside the docs directory, it
will attempt to build Sphinx docs as our own which will fail with a
confusing error.