fix: Generate data structures from Swagger Schema
In a previous changeset, I moved DataStructureGenerator to take a JSON
Schema instead of the Swagger Schema. This resulted in various
problems due to not fully supporting JSON Schema. This change reverts
back to the old behaviour of passing a Swagger Schema into
DataStructureGenerator and removing some of the handling of JSON Schema
specific properties (`examples`) will are not found in Swagger Schema.

As a side effect, this fixes a bug where generating a data structure
from a nullable type (`x-nullable`) was being converted to the `type`
array which wasn't supported. Thus now that the Swagger Schema is used
in DataStructure generator it receives `x-nullable` as expected.