test: Load chai extensions explicitly
The tests are currently relying on the fact that `test/unit/chai.js` is
loaded and affects global state of chai and thus the rest of the files
can access it.

Instead of relying on the extension changing global state of chai we
should explicitly load the extensions and use the chai module that it
returns.

This allows you to run test files individually, for example with editor
plugins. Otherwise the tests will fail because methods like `annotation`
on an Assertion does not exist:

    Error: Invalid Chai property: warning. Did you mean "string"?
      at Object.proxyGetter [as get] (node_modules/chai/lib/chai/utils/proxify.js:63:17)
      at Context.it (test/unit/parser/oas/parseSchemaObject-test.js:13:30)

    Error: Invalid Chai property: annotations
      at Object.proxyGetter [as get] (node_modules/chai/lib/chai/utils/proxify.js:66:17)
      at Context.it (test/unit/parser/oas/parseSchemaObject-test.js:25:36)