refactor(oas3): Schema Object parser should not accept member
The schema object parser shouldn't accept a member element with the ID
name because we can parse schema objects from other locations. Right now
this was only happening from components and thus there is always a
member with an ID.

The ID should be attached by the caller at use.