fix(oas2): prevent treating overlapping paths as a circular reference
A check that a reference to `#/definitions/User` from within
`#/definitions/UserList` would incorrectly been interpreted as a
circular reference as a simplistic check that the current path starts
with the current reference. We should ensure that the reference is
either identical, or relative from another (such as with a trailing
slash).